+++ Course File Management System/web/root/viewLecturers.jsp
@@ -8,6 +8,8 @@
 </head>
 <body>
     <div class="container">
+                
+        
       <table class="table">
         <caption>List of Lecturers</caption>
         <thead>
@@ -17,6 +19,7 @@
                 <th>Email Address</th>
                 <th>Phone Number</th>
                 <th>Status</th>
+                <th>Permission</th>
                 <th></th>
             </tr>
         </thead>
@@ -26,7 +29,7 @@
             ResultSet rs1 = DB.query(query1);
             while(rs1.next()) {
         %>
-        <tr><td><b>Dept. of <%=rs1.getString("department")%></b></td><td></td><td></td><td></td><td></td><td></td></tr>
+        <tr><td><b>Dept. of <%=rs1.getString("department")%></b></td><td></td><td></td><td></td><td></td><td></td><td></td></tr>
             <%
                 String query2 = "SELECT * FROM profile, user WHERE user.username = profile.username AND user.usertype = 'lecturer' AND profile.departmentID="+rs1.getString("departmentID");
                 ResultSet rs2 = DB.query(query2);   
@@ -38,13 +41,15 @@
                 <td><%=rs2.getString("emailAdd") %></td>
                 <td><%=rs2.getString("phoneNo") %></td>
                 <td><%=rs2.getString("status") %></td>
+                <td><%=rs2.getString("viewPERMISSION") %></td>
                 <td><a href="updateLecturer.jsp?username=<%=rs2.getString("username")%>"><button class="btn btn-primary btn-xs">Update</button></a>
-                <a href="<%=request.getContextPath()%>/deleteLecturer?username=<%=rs2.getString("username")%>"><button class="btn btn-danger btn-xs">Delete</button></a></td>
+                <a href="<%=request.getContextPath()%>/deleteLecturer?username=<%=rs2.getString("username")%>"><button class="btn btn-danger btn-xs" onclick= "return confirm('Are you sure you want to continue')">Delete</button></a></td>
             </tr>
             <% } %>
         <% } %>
         </tbody>
       </table>
+                
             <p>
    <a role="button" class="btn btn-primary" href="addLecturer.jsp">
       Add
